+++ fragmentino/graph.py
