+++ fragmentino/tests/medium_molecule_1.xyz
