--- framol/__init__.py
@@ -1,9 +1,0 @@
-from framol.molecule import Molecule
-from framol.io import FileHandlerXYZ
-from framol.graph import SimpleWeightedGraph
-from framol.graph import WeightedGraph
-from framol.molecular_fragmenter import MolecularFragmenter
-from framol.visualization_tools import Figure
-
-
-__version__ = "0.1"
